feat(gcp): add GCP project ID retrieval and integrate with metrics collection configuration
