+++ infra/metrics_collection.go
@@ -1,10 +1,7 @@
 package infra
 
 import (
-	"net/url"
 	"time"
-
-	"github.com/cockroachdb/errors"
 )
 
 type MetricCollectionConfig struct {
@@ -13,19 +10,24 @@ type MetricCollectionConfig struct {
 	MetricsIngestionURL string
 }
 
-// If metrics collection is enabled, the metrics ingestion url must be set
-func (cfg MetricCollectionConfig) Validate() error {
+// TODO: Before deploying in production, change the default url to the production url
+const DEFAULT_METRICS_INGESTION_URL = "http://localhost:8080/metrics"
+
+var PROJECT_ID_TO_URL = map[string]string{
+	"marble-prod-1":        "https://api.checkmarble.com/metrics",
+	"tokyo-country-381508": "https://api.staging.checkmarble.com/metrics",
+}
+
+// If metrics collection is enabled, build the metrics ingestion url from the project id
+func (cfg *MetricCollectionConfig) Configure() {
 	if !cfg.Enabled {
-		return nil
+		return
 	}
 
-	if cfg.MetricsIngestionURL == "" {
-		return errors.New("metrics ingestion url is not set")
+	// Build the MetricsIngestionURL from the project id
+	projectId, _ := GetProjectId()
+	cfg.MetricsIngestionURL = DEFAULT_METRICS_INGESTION_URL
+	if url, ok := PROJECT_ID_TO_URL[projectId]; ok {
+		cfg.MetricsIngestionURL = url
 	}
-
-	if _, err := url.ParseRequestURI(cfg.MetricsIngestionURL); err != nil {
-		return errors.Newf("invalid metrics ingestion url: %w", err)
-	}
-
-	return nil
 }
